LFS-795, LFS-366: Refactor Date questions and fix DateTime Split year, month and date questions into separate components
