+++ modules/data-entry/src/main/frontend/src/questionnaire/DateQuestionYear.jsx
@@ -1,0 +1,83 @@
+//
+//  Licensed to the Apache Software Foundation (ASF) under one
+//  or more contributor license agreements.  See the NOTICE file
+//  distributed with this work for additional information
+//  regarding copyright ownership.  The ASF licenses this file
+//  to you under the Apache License, Version 2.0 (the
+//  "License"); you may not use this file except in compliance
+//  with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing,
+//  software distributed under the License is distributed on an
+//  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+//  KIND, either express or implied.  See the License for the
+//  specific language governing permissions and limitations
+//  under the License.
+//
+
+import React from "react";
+
+import { withStyles } from "@material-ui/core";
+
+import PropTypes from "prop-types";
+
+import NumberQuestion from "./NumberQuestion";
+import QuestionnaireStyle from "./QuestionnaireStyle";
+
+import AnswerComponentManager from "./AnswerComponentManager";
+import DateQuestionUtilities from "./DateQuestionUtilities";
+
+// Component that renders a year only date question
+// Selected answers are placed in a series of <input type="hidden"> tags for
+// submission.
+//
+// Optional props:
+// text: the question to be displayed
+// type: "timestamp" for a single date or "interval" for two dates
+// dateFormat: yyyy
+// lowerLimit: lower date limit (inclusive) given as an object or string parsable by moment()
+// upperLimit: upper date limit (inclusive) given as an object or string parsable by moment()
+// Other options are passed to the <question> widget
+//
+// Sample usage:
+//<DateQuestionYear
+//  text="Please enter a date-time in 2019"
+//  dateFormat="yyyy-MM-dd hh:mm:ss"
+//  lowerLimit={new Date("01-01-2019")}
+//  upperLimit={new Date("12-31-2019")}
+//  type="timestamp"
+//  />
+function DateQuestionYear(props) {
+  let {existingAnswer, classes, ...rest} = props;
+  let {text, dateFormat, minAnswers, type, lowerLimit, upperLimit} = {dateFormat: "yyyy", minAnswers: 0, type: DateQuestionUtilities.TIMESTAMP_TYPE, ...props.questionDefinition, ...props};
+  return (
+    <NumberQuestion
+      minValue={0}
+      maxAnswers={1}
+      dataType="long"
+      errorText="Please insert a valid year."
+      isRange={(type === DateQuestionUtilities.INTERVAL_TYPE)}
+      answerNodeType="lfs:DateAnswer"
+      valueType="Long"
+      existingAnswer={existingAnswer}
+      maxValue={upperLimit || 9999}
+      minValue={lowerLimit || 0}
+      {...rest}
+      />
+  );
+}
+
+DateQuestionYear.propTypes = DateQuestionUtilities.PROP_TYPES;
+
+const StyledDateQuestion = withStyles(QuestionnaireStyle)(DateQuestionYear);
+export default StyledDateQuestion;
+
+AnswerComponentManager.registerAnswerComponent((questionDefinition) => {
+  if (questionDefinition.dataType === "date"
+    && DateQuestionUtilities.getDateType(questionDefinition.dateFormat) === DateQuestionUtilities.YEAR_DATE_TYPE)
+  {
+    return [StyledDateQuestion, 60];
+  }
+});
