+++ modules/data-entry/src/main/frontend/webpack.config.js
@@ -23,7 +23,7 @@ module.exports = {
     [module_name + 'subjectTypeIcon']: '@material-ui/icons/Category.js',
     [module_name + 'questionnairesIcon']: '@material-ui/icons/Assignment.js',
     [module_name + 'formsIcon']: '@material-ui/icons/Description.js',
-    [module_name + 'dateQuestion']: './src/questionnaire/DateQuestion.jsx',
+    [module_name + 'dateQuestionUtilities']: './src/questionnaire/DateQuestionUtilities.jsx',
     [module_name + 'userDashboard']: './src/dataHomepage/UserDashboard.jsx',
     [module_name + 'SubjectType']: './src/dataHomepage/SubjectType.jsx',
     [module_name + 'FormView']: './src/dataHomepage/FormView.jsx',
